+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## July 30, 2020
 - Modern JavaScript From The Beginning (Udemy)
+  - Section 2: JavaScript Language Fundamentals
+    - Variables - var, let & const
+- An example with var, let and const for Variables - var, let & const was added.
+
+## July 30, 2020
+- Modern JavaScript From The Beginning (Udemy)
   - Section 1: Intro & Getting Started
     - Project Files & Questions
     - Projects Github Repo